Age_and_Gender_Detection_requirements.txt
deepface
opencv-python
tensorflow
mysql-connector-python